% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ring_app_launch.R
\name{ring_app_launch}
\alias{ring_app_launch}
\title{Run Shiny-based Application}
\usage{
ring_app_launch(launch.browser = TRUE)
}
\arguments{
\item{launch.browser}{A logical value. If \code{TRUE}, the system's default 
web browser will be launched automatically after the app is started.}
}
\description{
Run a Shiny-based application within the system's default 
web browser.
}
\details{
A workflow for the Shiny app can be found in the README file of this 
package. Here is a link \url{https://github.com/JingningShi/MtreeRing}

To stop the app, press the Escape key, or click the stop sign icon 
in the upper right corner of the RStudio console.
}
\author{
Jingning Shi, Wei Xiang
}
