% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchMtRApp.R
\name{launchMtRApp}
\alias{launchMtRApp}
\title{Run Shiny-based Application}
\usage{
launchMtRApp()
}
\description{
Run a Shiny-based application within the system's default web browser.
}
\details{
A workflow for the Shiny app can be found in the README file of this package. Here is a link: https://github.com/JingningShi/MtreeRing
}
\author{
Jingning Shi, Wei Xiang
}
