\name{launchMtRApp}
\alias{launchMtRApp}
\title{Run Shiny-based Application}
\description{Run a Shiny-based application within the system's default web browser. The application provides a beginner-friendly graphical interface and supports more flexible mouse-based interactions.

A workflow for the Shiny app can be found in the README file of this package. Here is a link: https://github.com/JingningShi/MtreeRing.}
\usage{launchMtRApp()}
\author{Jingning Shi, Wei Xiang}
