\name{MtreeRing-package}
\alias{MtreeRing-package}
\alias{MtreeRing}
\docType{package}
\title{A Shiny Application for Automatic Measurements of Tree-Ring Widths on Digital Images}
\description{
Use morphological image processing and edge detection algorithms to automatically identify tree-ring boundaries on digital images. Tree-ring boundaries are determined by the changes of light reflectance from late wood to early wood. Two geometric models are provided to calibrate the errors resulted from inclined rings. The package provides a Shiny-based application, allowing R beginners to easily analyze tree-ring images and export ring-width series in standard file formats.
}
\details{

\tabular{ll}{
Package: \tab MtreeRing\cr
Type: \tab Package\cr
License: \tab GPL-3\cr
Maintainer: \tab Jingning Shi <snow940220@bjfu.edu.cn>\cr
LazyData: \tab TRUE\cr
}

}
\author{Jingning Shi, Wei Xiang}
\keyword{ package }