% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ring_app_launch.R
\name{ring_app_launch}
\alias{ring_app_launch}
\title{Run Shiny-based Application}
\usage{
ring_app_launch(launch.browser = TRUE)
}
\arguments{
\item{launch.browser}{A logical value. If \code{TRUE}, the system's default 
web browser will be launched automatically after the app is started.}
}
\description{
Run a Shiny-based application within the system's default 
web browser.
}
\details{
A workflow for the Shiny app can be found in the package vignette. 
Type \code{vignette('app-MtreeRing')} to learn how to use the app. Most
steps are demonstrated with a gif to make the workflow more understandable.

To stop the app, go to the R console and press the Escape key. 
You can also click the stop sign icon in the
upper right corner of the RStudio console.
}
\author{
Jingning Shi, Wei Xiang
}
